ST-13: Move some tests now that settings are accessed from homePage
