+++ test/home/view/goldens/start_game_form.phone.png
